Separated HeroCarousel from TopNav
