+++ src/component/TopNavComponent.js
@@ -32,40 +32,6 @@ class TopNav extends Component {
 	}
 
 	render() {
-		const RenderCarousel = ({ drinks }) => {
-			// const checkFeatured = (drink) => drink.featured === true
-			const slides = [
-				{
-					src: drinks[0].image,
-					altText: drinks[0].name,
-					caption: drinks[0].description,
-					header: drinks[0].name,
-					key: '1',
-				},
-				{
-					src: drinks[1].image,
-					altText: drinks[1].name,
-					caption: drinks[1].description,
-					header: drinks[1].name,
-					key: '2',
-				},
-				{
-					src: drinks[2].image,
-					altText: drinks[2].name,
-					caption: drinks[2].description,
-					header: drinks[2].name,
-					key: '3',
-				},
-			]
-
-			console.log(slides)
-
-			return (
-				<div className="carouselImage">
-					<UncontrolledCarousel items={slides} />
-				</div>
-			)
-		}
 
 		return (
 			<React.Fragment>
@@ -105,7 +71,6 @@ class TopNav extends Component {
 					<LoginModal />
 					<SearchModal />
 				</Navbar>
-					<RenderCarousel drinks={this.props.drinks} />
 			</React.Fragment>
 		)
 	}
